requirements.txt
pyspark==3.4.0
azure-storage-blob==12.19.0
azure-identity==1.15.0
databricks-cli==0.18.0
python-dotenv==1.0.0
requests==2.31.0